+++ packages/injective-notifications/src/module.ts
@@ -11,11 +11,11 @@ export default defineNuxtModule({
       dirs.push({
         path: fileURLToPath(new URL('./../lib/components', import.meta.url))
       })
-    }
+    },
 
-    // 'imports:dirs'(dirs) {
-    //   dirs.push(fileURLToPath(new URL('./../lib/composables', import.meta.url)))
-    // }
+    'imports:dirs'(dirs) {
+      dirs.push(fileURLToPath(new URL('./../lib/composables', import.meta.url)))
+    }
 
     // 'imports:sources'(sources) {
     //   sources.push({
